Reorganized some views and forms
--- `checkin` ---

Moved from `views.py` to `api/views.py`:
* AdminAPISuggestSkillVoteView
* AdminAPISuggestSkillDeleteView
* AdminAPIRegisterProfileView

--- `make_queue` ---

Moved from `forms.py` to `api/forms.py`:
* APIMachineDataQueryForm
* APIReservationListQueryForm
Moved from `forms.py` to `forms/course.py`:
* Printer3DCourseForm
Moved from `forms.py` to `forms/machine.py`:
* MachineFormBase
* AddMachineForm
* ChangeMachineForm
* MachineDetailQueryForm
Moved from `forms.py` to `forms/quota.py`:
* QuotaForm
* AdminQuotaPanelQueryForm
Moved from `forms.py` to `forms/reservation_rule.py`:
* ReservationRuleForm
Moved from `forms.py` to `forms/reservation.py`:
* ReservationForm
* ReservationFindFreeSlotsForm
* ReservationListQueryForm

Moved from `views/api/calendar.py` to `api/views.py`:
* reservation_type()
* APIReservationListView
* APIReservationRuleListView
Moved from `views/api/reservation.py` to `api/views.py`:
* APIMachineDataView
Moved from `views/reservation/reservation.py` to `api/views.py`:
* APIReservationMarkFinishedView
* APIReservationDeleteView

Renamed `views/admin/course.py` to `views/course.py`.

Moved from `views/reservation/rules.py` to `views/machine.py`:
* MachineTypeRelatedViewMixin
* MachineUsageRuleDetailView
* MachineUsageRuleUpdateView
Moved from `views/reservation/reservation.py` to `views/machine.py`:
* MachineRelatedViewMixin
Moved from `views/reservation/calendar.py` to `views/machine.py`:
* MachineDetailView
Moved from `views/reservation/machine.py` to `views/machine.py`:
* MachineListView
* MachineFormMixin
* MachineCreateView
* MachineUpdateView
* MachineDeleteView

Renamed `views/admin/quota.py` to `views/quota.py`.
Moved from `views/quota/user.py` to `views/quota.py`:
* AdminUserQuotaListView

Moved from `views/reservation/rules.py` to `views/reservation_rule.py`:
* ReservationRuleListView
* ReservationRuleFormMixin
* ReservationRuleCreateView
* ReservationRuleUpdateView
* ReservationRuleDeleteView

Moved from `views/reservation/reservation.py` to `views/reservation.py`:
* ReservationCreateOrUpdateView
* ReservationCreateView
* ReservationUpdateView
* ReservationListView
* ReservationFindFreeSlotsView

--- `news` ---

Moved from `views.py` to `api/views.py`:
* AdminAPINewsBaseToggleView
* AdminAPIArticleToggleView
* AdminAPIEventToggleView
* AdminAPITimePlaceToggleView
Moved from `views.py` to `views/article.py`:
* ArticleListView
* ArticleDetailView
* AdminArticleListView
* NewsBaseFormMixin
* ArticleFormMixin
* ArticleUpdateView
* ArticleCreateView
* ArticleDeleteView
Moved from `views.py` to `views/event.py`:
* EventListView
* EventDetailView
* AdminEventListView
* AdminEventParticipantsSearchView
* AdminEventDetailView
* EventFormMixin
* EventUpdateView
* EventCreateView
* EventRelatedViewMixin
* TimePlaceRelatedViewMixin
* TimePlaceFormMixin
* TimePlaceUpdateView
* TimePlaceCreateView
* TimePlaceDuplicateCreateView
* EventDeleteView
* TimePlaceDeleteView
* EventTicketCreateView
* EventTicketDetailView
* EventTicketMyListView
* AdminEventTicketListView
* AdminTimePlaceTicketListView
* EventTicketCancelView

--- `users` ---

Renamed `views.py` to `api/views.py`.
