+++ src/make_queue/tests/forms/test_machine_forms.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 
 from util.locale_utils import parse_datetime_localized
 from util.test_utils import with_time
-from ...forms import AddMachineForm, ChangeMachineForm
+from ...forms.machine import AddMachineForm, ChangeMachineForm
 from ...models.machine import Machine, MachineType
