+++ src/make_queue/tests/views/test_api_reservation_views.py
@@ -10,10 +10,10 @@ from django_hosts import reverse
 from news.models import Event, TimePlace
 from users.models import User
 from util.locale_utils import parse_datetime_localized
+from ...api.views import APIReservationListView
 from ...models.course import Printer3DCourse
 from ...models.machine import Machine, MachineType
 from ...models.reservation import Quota, Reservation
-from ...views.api.calendar import APIReservationListView
 
 
 class APIReservationListViewTests(TestCase):
